+++ bathbot/src/active/impls/profile/top100_stats.rs
@@ -91,7 +91,7 @@ impl Top100Stats {
             let map_attrs = map
                 .attributes()
                 .mods(score.mods.bits())
-                .clock_rate(score.mods.clock_rate().unwrap_or(1.0) as f64)
+                .clock_rate(score.mods.clock_rate().unwrap_or(1.0))
                 .build();
 
             this.ar.add(map_attrs.ar);
